+++ crates/remote-prover-client/src/remote_prover/batch_prover.rs
@@ -31,6 +31,7 @@ pub struct RemoteBatchProver {
     client: Arc<Mutex<Option<ApiClient<tonic::transport::Channel>>>>,
 
     endpoint: String,
+    timeout: Duration,
 }
 
 impl RemoteBatchProver {
@@ -40,7 +41,23 @@ impl RemoteBatchProver {
         RemoteBatchProver {
             endpoint: endpoint.into(),
             client: Arc::new(Mutex::new(None)),
+            timeout: Duration::from_secs(10),
         }
+    }
+
+    /// Configures the timeout for requests to the remote prover server.
+    ///
+    /// # Arguments
+    ///
+    /// * `timeout` - The timeout duration for requests.
+    ///
+    /// # Returns
+    ///
+    /// A new [`RemoteBatchProver`] instance with the configured timeout.
+    #[must_use]
+    pub fn with_timeout(mut self, timeout: Duration) -> Self {
+        self.timeout = timeout;
+        self
     }
 
     /// Establishes a connection to the remote batch prover server. The connection is
@@ -62,7 +79,7 @@ impl RemoteBatchProver {
         let new_client = {
             let endpoint = tonic::transport::Endpoint::try_from(self.endpoint.clone())
                 .map_err(|err| RemoteProverClientError::ConnectionFailed(err.into()))?
-                .timeout(Duration::from_millis(10000));
+                .timeout(self.timeout);
             let channel = endpoint
                 .tls_config(tonic::transport::ClientTlsConfig::new().with_native_roots())
                 .map_err(|err| RemoteProverClientError::ConnectionFailed(err.into()))?